Remove cache file when run Prettier without `--cache` option
