+++ src/cli/utils.js
@@ -1,5 +1,7 @@
 "use strict";
 
+const { promises: fs } = require("fs");
+
 // eslint-disable-next-line no-restricted-modules
 const { default: sdbm } = require("../../vendors/sdbm.js");
 
@@ -49,4 +51,20 @@ function createHash(source) {
   return sdbm(source).toString();
 }
 
-module.exports = { printToScreen, groupBy, pick, createHash };
+/**
+ * Get stats of a given path.
+ * @param {string} filePath The path to target file.
+ * @returns {Promise<import('fs').Stats | undefined>} The stats.
+ */
+async function statSafe(filePath) {
+  try {
+    return await fs.stat(filePath);
+  } catch (error) {
+    /* istanbul ignore next */
+    if (error.code !== "ENOENT") {
+      throw error;
+    }
+  }
+}
+
+module.exports = { printToScreen, groupBy, pick, createHash, statSafe };
